requirements.txt
# Web 框架
Flask==2.0.1
Flask-CORS==3.0.10

# 数据处理
requests==2.28.1
numpy==1.23.4

# 环境变量
python-dotenv==0.20.0

# Qdrant 客户端
qdrant-client==1.16.0

# 嵌入模型
llama-index-embeddings-dashscope==0.1.0
llama-index-core==0.1.0

# PDF 处理
pymupdf==1.21.0

# 其他
jinja2==3.0.3
werkzeug==2.0.3
